Finish observable layer that holds all the common logic for each observable type
