+++ library/src/main/java/com/anthonycr/bonsai/CompletableSubscriberWrapper.java
@@ -29,40 +29,20 @@ import android.support.annotation.Nullable;
  * the callbacks on the correct threads, and throws the
  * appropriate errors when certain rules are violated.
  */
-class CompletableSubscriberWrapper<T extends CompletableOnSubscribe> implements CompletableSubscriber {
+@SuppressWarnings("WeakerAccess")
+class CompletableSubscriberWrapper<T extends CompletableOnSubscribe> extends ObservableSubscriberWrapper<T> implements CompletableSubscriber {
 
-    private volatile boolean onStartExecuted = false;
-    volatile boolean onCompleteExecuted = false;
-    volatile boolean onErrorExecuted = false;
-
-    @Nullable private final Scheduler observerThread;
-    @NonNull private final Scheduler defaultThread;
-    @Nullable protected volatile T onSubscribe;
+    protected volatile boolean onCompleteExecuted;
 
     CompletableSubscriberWrapper(@Nullable T onSubscribe,
                                  @Nullable Scheduler observerThread,
                                  @NonNull Scheduler defaultThread) {
-        this.onSubscribe = onSubscribe;
-        this.observerThread = observerThread;
-        this.defaultThread = defaultThread;
-    }
-
-    void executeOnObserverThread(@NonNull Runnable runnable) {
-        if (observerThread != null) {
-            observerThread.execute(runnable);
-        } else {
-            defaultThread.execute(runnable);
-        }
-    }
-
-    @Override
-    public void unsubscribe() {
-        onSubscribe = null;
+        super(onSubscribe, observerThread, defaultThread);
     }
 
     @Override
     public void onComplete() {
-        CompletableOnSubscribe onSubscribe = this.onSubscribe;
+        T onSubscribe = this.onSubscribe;
 
         if (onCompleteExecuted) {
             throw new RuntimeException("onComplete called more than once");
@@ -73,36 +53,5 @@ class CompletableSubscriberWrapper<T extends CompletableOnSubscribe> implements 
         onCompleteExecuted = true;
 
         unsubscribe();
-    }
-
-    @Override
-    public void onStart() {
-        CompletableOnSubscribe onSubscribe = this.onSubscribe;
-
-        if (onStartExecuted) {
-            throw new RuntimeException("onStart is called internally, do not call it yourself");
-        } else if (onSubscribe != null) {
-            executeOnObserverThread(new OnStartRunnable(onSubscribe));
-        }
-
-        onStartExecuted = true;
-    }
-
-    @Override
-    public void onError(@NonNull final Throwable throwable) {
-        CompletableOnSubscribe onSubscribe = this.onSubscribe;
-
-        if (onSubscribe != null) {
-            executeOnObserverThread(new OnErrorRunnable(onSubscribe, throwable));
-        }
-
-        onErrorExecuted = true;
-
-        unsubscribe();
-    }
-
-    @Override
-    public boolean isUnsubscribed() {
-        return onSubscribe == null;
     }
 }
